refactor: sidebar refactor
