+++ src/app/(payload)/admin/importMap.js
@@ -65,7 +65,7 @@ import { default as default_1a7510af427896d367a49dbf838d2de6 } from '@/component
 import { default as default_8a7ab0eb7ab5c511aba12e68480bfe5e } from '@/components/BeforeLogin'
 import { default as default_78b39f784b86cf844383560b5f1a85bf } from '@/components/ColorPicker/'
 import { default as default_545a25a7f6f58e32e2350bd4892658ab } from '@/components/IconSelect/'
-import { default as default_ecb58de83adb754b6c9d0ac4f2801b7b } from '@/components/Nav/WordpressMigration/Nav'
+import { default as default_aeea2462fad41415a04ffbf0a5f77d84 } from '@/components/Nav/WordpressMigration'
 import { DefaultRootView as DefaultRootView_a5b063f5b2b9cdce4f0e1dd9b2761872 } from '@/components/views/WordpressMigration/DefaultRootView'
 import { AIFeature as AIFeature_c86cb9794e8dfc789dceffc1ecbdea8d } from '@/fields/ai/client'
 import { SlugComponent as SlugComponent_92cc057d0a2abb4f6cf0307edf59f986 } from '@/fields/slug/SlugComponent'
@@ -145,7 +145,7 @@ export const importMap = {
   '@payloadcms/plugin-search/client#ReindexButton': ReindexButton_aead06e4cbf6b2620c5c51c9ab283634,
   '@payloadcms/plugin-multi-tenant/rsc#GlobalViewRedirect':
     GlobalViewRedirect_d6d5f193a167989e2ee7d14202901e62,
-  '@/components/Nav/WordpressMigration/Nav#default': default_ecb58de83adb754b6c9d0ac4f2801b7b,
+  '@/components/Nav/WordpressMigration#default': default_aeea2462fad41415a04ffbf0a5f77d84,
   '@/components/BeforeDashboard#default': default_1a7510af427896d367a49dbf838d2de6,
   '@/components/BeforeLogin#default': default_8a7ab0eb7ab5c511aba12e68480bfe5e,
   '@payloadcms/plugin-multi-tenant/client#TenantSelector':
